Add shaper abstractions
